Installation de l'authentification

+++ app/Http/Controllers/CheckoutController.php
@@ -2,11 +2,14 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
-use Stripe\Stripe;
+use DateTime;
+use App\Models\Order;
 use Stripe\PaymentIntent;
-use Gloudemans\Shoppingcart\Facades\Cart;
+use Stripe\Stripe;
 use Illuminate\Support\Arr;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Session;
+use Gloudemans\Shoppingcart\Facades\Cart;
 
 class CheckoutController extends Controller
 {
@@ -17,16 +20,15 @@ class CheckoutController extends Controller
      */
     public function index()
     {
-
         if (Cart::count() <= 0) {
             return redirect()->route('products.index');
         }
 
-        Stripe::setApiKey('sk_test_51Hxb1GF5feTRJ8zxA2GOvUCQ7gM6WHVEqpWXZn7x0HwYSnq2yn2jvK9iDv26OKHHud00cWXPaQP1mqeFoPevH1fm00xzzzpd7q');
+        Stripe::setApiKey('sk_test_3WteeitM6Wi4AK3SdJzBrm7300qGrAamxX');
 
         $intent = PaymentIntent::create([
-          'amount' => round(Cart::total()),
-          'currency' => 'eur',
+            'amount' => round(Cart::total()),
+            'currency' => 'eur'
         ]);
 
         $clientSecret = Arr::get($intent, 'client_secret');
@@ -54,11 +56,43 @@ class CheckoutController extends Controller
      */
     public function store(Request $request)
     {
-        Cart::destroy();
-        
         $data = $request->json()->all();
 
-        return $data['paymentIntent'];
+        $order = new Order();
+
+        $order->payment_intent_id = $data['paymentIntent']['id'];
+        $order->amount = $data['paymentIntent']['amount'];
+
+        $order->payment_created_at = (new DateTime())
+            ->setTimestamp($data['paymentIntent']['created'])
+            ->format('Y-m-d H:i:s');
+
+        $products = [];
+        $i = 0;
+
+        foreach (Cart::content() as $product) {
+            $products['product_' . $i][] = $product->model->title;
+            $products['product_' . $i][] = $product->model->price;
+            $products['product_' . $i][] = $product->qty;
+            $i++;
+        }
+
+        $order->products = serialize($products);
+        $order->user_id = Auth()->user()->id;
+        $order->save();
+
+        if ($data['paymentIntent']['status'] === 'succeeded') {
+            Cart::destroy();
+            Session::flash('success', 'Votre commande a été traitée avec succès.');
+            return response()->json(['success' => 'Payment Intent Succeeded']);
+        } else {
+            return response()->json(['error' => 'Payment Intent Not Succeeded']);
+        }
+    }
+
+    public function thankyou()
+    {
+        return Session::has('success') ? view('checkout.thankYou') : redirect()->route('products.index');
     }
 
     /**
@@ -104,10 +138,5 @@ class CheckoutController extends Controller
     public function destroy($id)
     {
         //
-    }
-
-    public function thankyou() {
-        
-        return view('checkout.thankyou');
     }
 }
